fix: correct id type

+++ src/main/java/com/navangs/maribong/repository/HistoryRepository.java
@@ -3,6 +3,6 @@ package com.navangs.maribong.repository;
 import com.navangs.maribong.dao.History;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HistoryRepository extends JpaRepository<History, Long> {
+public interface HistoryRepository extends JpaRepository<History, Integer> {
     History findByUserId(String userId);
 }
